refactor: CustomerHeader
