+++ src/partials/customer/header/components/header-menu/header-menu.module.css
@@ -1,0 +1,238 @@
+.header__menu {
+  width: min(100%, 1152px);
+  padding-inline: 1rem;
+  display: flex;
+  justify-content: flex-start;
+  position: relative;
+}
+
+.menu__item {
+  height: 3rem;
+  padding: .75rem;
+
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: .25rem;
+
+  cursor: pointer;
+  text-overflow: ellipsis;
+  overflow: hidden;
+  white-space: nowrap;
+}
+
+.menu__item .menu__cate {
+  display: flex;
+  align-items: center;
+  gap: .25rem;
+}
+
+.menu__item .menu__cate-dropdown {
+  position: absolute;
+  left: 0;
+  right: 0;
+  top: 100%;
+
+  width: 100%;
+  max-width: 1152px;
+  min-height: 494px;
+  padding: .75rem;
+  border-radius: 0 0 .75rem .75rem;
+
+  display: none;
+  flex-direction: column;
+  justify-content: flex-start;
+
+  background-color: var(--white);
+  box-shadow: 0px .5rem .5rem var(--black-o20);
+
+  cursor: default;
+}
+
+.cate-dropdown__info {
+  width: calc((100% - .75rem) * .25 - .75rem);
+  border-radius: .25rem;
+
+  display: flex;
+  justify-content: flex-start;
+  align-items: center;
+  gap: .25rem;
+  padding: .25rem .75rem;
+
+}
+
+.cate-dropdown__info .cate-dropdown__img {
+  width: 2.5rem;
+  height: 2.5rem;
+  padding: .25rem;
+  object-fit: contain;
+}
+
+.cate-dropdown__content {
+  width: calc((100% - .75rem) * .75 - .75rem);
+  position: absolute;
+  top: .75rem;
+  right: .75rem;
+  bottom: .75rem;
+
+  /* padding: 0 1.5rem; */
+
+  display: none;
+  flex-direction: column;
+  gap: .75rem;
+}
+
+.cate-dropdown__title {
+  height: 3rem;
+
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+
+  color: var(--primary-color-2);
+}
+
+.cate-dropdown__title-link {
+  text-decoration: none;
+  display: flex;
+  color: inherit;
+  gap: .25rem;
+}
+
+.cate-dropdown__title-link:hover {
+  font-weight: 500;
+}
+
+.cate-dropdown__products {
+  display: grid;
+  grid-template-columns: repeat(4, 1fr);
+  gap: .75rem;
+}
+
+.cate-dropdown__product-img {
+  width: 100px;
+  height: 100px;
+  object-fit: contain;
+}
+
+.cate-dropdown__product-link {
+  padding: .5rem;
+
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 4px;
+
+  color: var(--black);
+  text-decoration: none;
+}
+
+.cate-dropdown__product-link:hover {
+  box-shadow: inset 0px 0px 0px 1px var(--primary-color-2);
+  border-radius: .5rem;
+  background-color: var(--bg-gray);
+}
+
+.cate-dropdown__product-price {
+  font-weight: 500;
+
+}
+
+.cate-dropdown__product-price--discount {
+  font-weight: 400;
+  color: var(--del-gray);
+}
+
+.cate-dropdown__product-name {
+  text-overflow: ellipsis;
+  overflow: hidden;
+  white-space: nowrap;
+}
+
+.cate-dropdown__product-rating {
+  display: flex;
+  justify-content: flex-start;
+
+}
+
+.cate-dropdown__product-star {
+  font-size: 1.25rem;
+  color: var(--primary-color-5);
+}
+
+.cate-dropdown__products--not-found {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: .75rem;
+}
+
+.cate-dropdown__products--not-found__img {
+  width: 10rem;
+  height: 16rem;
+  object-fit: contain;
+}
+
+.cate-dropdown__products--not-found__text {
+  font-size: 1.25rem;
+  font-weight: 500;
+  color: var(--primary-color-2);
+}
+
+.menu__item .cate-dropdown__wrapper:last-child {
+  order: -1;
+}
+
+.menu__item .cate-dropdown__wrapper:last-child .cate-dropdown__info {
+  background-color: var(--bg-white-o80);
+}
+
+.menu__item .cate-dropdown__wrapper:last-child .cate-dropdown__content {
+  display: flex;
+}
+
+.menu__item .menu__hot-product {
+  color: var(--fire-orange);
+}
+
+.menu__item .menu__promo {
+  color: var(--pink);
+}
+
+.menu__item:hover {
+  color: var(--primary-color-2);
+  box-shadow: inset 0 -2px 0 0 var(--primary-color-2);
+
+}
+
+.menu__item:hover .menu__cate-dropdown {
+  display: flex;
+}
+
+.menu__item:hover .cate-dropdown__wrapper:hover .cate-dropdown__info {
+  background-color: var(--bg-white-o80);
+}
+
+.menu__item:hover .cate-dropdown__wrapper:hover .cate-dropdown__content {
+  display: flex;
+}
+
+.menu__item:hover .cate-dropdown__wrapper:hover~.cate-dropdown__wrapper .cate-dropdown__info {
+  background-color: transparent;
+}
+
+.menu__item:hover .cate-dropdown__wrapper:hover~.cate-dropdown__wrapper .cate-dropdown__content {
+  display: none;
+}
+
+.header__menu a {
+  text-decoration: none;
+  color: var(--black);
+}
+
+.active-menu-item {
+  color: var(--primary-color-2) !important;
+  box-shadow: inset 0 -2px 0 0 var(--primary-color-2);
+}
