+++ src/partials/customer/header/components/header-nav/header-nav.module.css
@@ -1,0 +1,268 @@
+.header__nav {
+  width: 100%;
+  height: 2.5rem;
+  margin-bottom: 1rem;
+
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  font-size: 0.8125rem;
+  background-color: var(--primary-color-1);
+}
+
+.header__nav .material-symbols-outlined {
+  font-size: 1rem;
+  /* margin-right: .3rem; */
+}
+
+.header__nav .help--disable,
+.header__nav .noti--disable {
+  display: none;
+  align-items: center;
+}
+
+.header__nav-container {
+  width: 100%;
+  padding-inline: 1rem;
+  max-width: 1152px;
+  display: flex;
+  justify-content: space-between;
+
+}
+
+.header__nav-container .header__support-info {
+  display: flex;
+  gap: 1rem;
+}
+
+.header__nav--disable {
+  display: none;
+  justify-content: flex-end;
+  gap: .5rem;
+}
+
+.header__nav--disable .material-symbols-outlined {
+  font-size: 1.5rem;
+}
+
+.header__nav a {
+  color: white;
+  margin: auto;
+  display: flex;
+  align-items: center;
+  gap: .25rem;
+  box-sizing: border-box;
+}
+
+.dropdown-help {
+  position: relative;
+  display: inline-block;
+}
+
+.dropdown-help__content-container {
+  display: none;
+  position: absolute;
+  z-index: 1;
+  width: 17.5rem;
+  border-radius: 0 0 .75rem .75rem;
+  padding-top: .5rem;
+}
+
+.dropdown-help__content {
+  background-color: #f9f9f9;
+  box-shadow: 0 .5rem 1rem var(--black-o20);
+  z-index: 1;
+  padding: .75rem;
+  flex-direction: column;
+  align-items: center;
+  width: 100%;
+  border-radius: 0 0 .5rem .5rem;
+}
+
+.dropdown-help__content img {
+  object-fit: cover;
+  width: 100%;
+}
+
+.dropdown-help__content p {
+  text-align: center;
+  font-size: 0.875rem;
+  font-weight: 500;
+}
+
+.dropdown-help:hover .dropdown-help__content-container {
+  display: flex;
+}
+
+.dropdown-help:hover a {
+  color: #d7d7d7;
+}
+
+.header__about-account {
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+}
+
+.dropdown-noti {
+  position: relative;
+
+  display: inline-flex;
+}
+
+.dropdown-noti__content-container {
+  width: 20rem;
+  position: absolute;
+  top: 100%;
+  right: -.5rem;
+  z-index: 1;
+
+  padding-top: .5rem;
+  display: none;
+
+  border-radius: 0 0 .5rem .5rem;
+}
+
+.dropdown-noti__content {
+  width: 100%;
+  z-index: 1;
+
+  border-radius: 0 0 .75rem .75rem;
+
+  flex-direction: column;
+  align-items: center;
+
+  background-color: #f9f9f9;
+  box-shadow: 0 .5rem 1rem var(--black-o20);
+}
+
+.dropdown-noti__content .dropdown-noti__unauth-user {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding: 3.5rem;
+}
+
+.dropdown-noti__content .dropdown-noti__content-title {
+  color: var(--dark7);
+  display: flex;
+  margin-bottom: .75rem;
+  margin-top: .75rem;
+  margin-left: .5rem;
+}
+
+.dropdown-noti__content-container .dropdown-noti__btn-read-all {
+  display: flex;
+  justify-content: center;
+  text-align: center;
+  text-decoration: none;
+  color: var(--primary-color-1);
+  margin: .75rem auto;
+  font-weight: 500;
+}
+
+.dropdown-noti__content-container .dropdown-noti__btn-read-all:hover {
+  color: var(--primary-color-2);
+}
+
+.dropdown-noti:hover .dropdown-noti__content-container {
+  display: flex;
+}
+
+.dropdown-noti:hover .header__notifications {
+  color: #d7d7d7;
+}
+
+.unauth-user__img {
+  object-fit: contain;
+  width: 5rem;
+}
+
+.unauth-content__noti {
+  font-size: 0.875rem;
+  font-weight: 400;
+  color: black;
+}
+
+.unauth-content__btn {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: .5rem;
+  background-color: var(--outline-gray);
+  border-radius: 0 0 .75rem .75rem;
+}
+
+.unauth-content__btn a {
+  text-decoration: none;
+  text-align: center;
+  color: black;
+  font-size: 0.875rem;
+  margin: 0 auto;
+}
+
+.unauth-content__btn a:hover {
+  color: var(--primary-color-1);
+}
+
+.dropdown-noti--empty {
+  display: flex;
+  flex-direction: column;
+  padding: 62px;
+  align-items: center;
+  gap: .5rem;
+}
+
+.dropdown-noti__img--empty {
+  object-fit: contain;
+  width: 100px;
+}
+
+.empty-content__noti {
+  color: black;
+  font-size: 0.875rem
+}
+
+.header__support-info__hotline:hover {
+  color: var(--dark11);
+}
+
+.header__logo--top {
+  display: none;
+}
+
+.header__logo--top img {
+  object-fit: contain;
+  width: 10rem;
+}
+
+.header__about-account .header__notifications {
+  display: flex;
+}
+
+.header__auth {
+  display: flex;
+  color: white;
+  gap: 4px;
+  align-items: center;
+}
+
+.header__auth:hover {
+  cursor: pointer;
+  color: var(--dark11);
+}
+
+.header__auth img {
+  object-fit: cover;
+  height: 30px;
+  border-radius: 50%;
+}
+
+.header__auth-login:hover {
+  color: #d7d7d7;
+}
+
+.header__auth-register:hover {
+  color: #d7d7d7;
+}
