+++ src/partials/customer/header/components/index.tsx
@@ -1,0 +1,5 @@
+import HeaderNav from "./header-nav";
+import HeaderMain from "./header-main";
+import HeaderMenu from "./header-menu";
+
+export { HeaderNav, HeaderMain, HeaderMenu };
